requirements.txt
flask
python-dotenv
langchain-community
langchain-openai
faiss-cpu
openai
tiktoken